feat(ui-core): extract shared project UI components to ui-core (RHOAIENG-65099 Phase 1)
Move multi-consumer design system components from @odh-dashboard/internal
to @odh-dashboard/ui-core: CollapsibleSection, EmptyDetailsView,
ProjectObjectType, SectionType, color functions, TypeBorderedCard,
WhosMyAdministrator, PopoverListContent, and vars.scss CSS custom properties.

Update all package consumers (model-serving, autorag, automl, maas,
feature-store, eval-hub, mlflow-embedded, agent-ops, kserve,
model-serving-backport, model-training) to import from ui-core.
Re-export shims in frontend/src/ for backward compatibility.
