+++ packages/autorag/frontend/src/app/components/empty-states/EmptyExperimentsState.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Link } from 'react-router';
 
-import EmptyDetailsView from '@odh-dashboard/internal/components/EmptyDetailsView';
-import { ProjectObjectType, typedEmptyImage } from '@odh-dashboard/internal/concepts/design/utils';
+import { EmptyDetailsView, ProjectObjectType } from '@odh-dashboard/ui-core';
+import { typedEmptyImage } from '@odh-dashboard/internal/concepts/design/utils';
 import { Button } from '@patternfly/react-core';
 
 /**
