+++ packages/feature-store/src/FeatureStoreCoreLoader.tsx
@@ -7,7 +7,7 @@ import { conditionalArea } from '@odh-dashboard/internal/concepts/areas/AreaComp
 import { SupportedArea } from '@odh-dashboard/plugin-core/areas';
 // eslint-disable-next-line @odh-dashboard/no-restricted-imports
 import ApplicationsPage from '@odh-dashboard/internal/pages/ApplicationsPage';
-import WhosMyAdministrator from '@odh-dashboard/internal/components/WhosMyAdministrator';
+import { WhosMyAdministrator } from '@odh-dashboard/ui-core';
 // eslint-disable-next-line @odh-dashboard/no-restricted-imports
 import RedirectErrorState from '@odh-dashboard/internal/pages/external/RedirectErrorState';
 import { useAccessAllowed } from '@odh-dashboard/internal/concepts/userSSAR/useAccessAllowed';
