+++ packages/kserve/extensions.ts
@@ -1,7 +1,6 @@
 // eslint-disable-next-line no-restricted-syntax, @odh-dashboard/no-restricted-imports
 import { NamespaceApplicationCase } from '@odh-dashboard/internal/pages/projects/types';
-// eslint-disable-next-line no-restricted-syntax
-import { ProjectObjectType } from '@odh-dashboard/internal/concepts/design/utils';
+import type { ProjectObjectType } from '@odh-dashboard/ui-core';
 import type {
   ModelServingPlatformExtension,
   ModelServingDeleteModal,
