+++ packages/maas/frontend/src/app/pages/keys-and-subs/ApiKeysAndSubscriptionsPage.tsx
@@ -1,6 +1,6 @@
 import ApplicationsPage from '@odh-dashboard/internal/pages/ApplicationsPage';
 import TitleWithIcon from '@odh-dashboard/internal/concepts/design/TitleWithIcon';
-import { ProjectObjectType } from '@odh-dashboard/internal/concepts/design/utils';
+import { ProjectObjectType } from '@odh-dashboard/ui-core';
 import { Tab, Tabs, TabTitleText } from '@patternfly/react-core';
 import React from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
