+++ packages/mlflow-embedded/experiments/GlobalMLflowExperimentsRoutes.tsx
@@ -11,7 +11,7 @@ import { Navigate, Route } from 'react-router-dom';
 import NotFound from '@odh-dashboard/internal/pages/NotFound';
 import ProjectsRoutes from '@odh-dashboard/internal/concepts/projects/ProjectsRoutes';
 import TitleWithIcon from '@odh-dashboard/internal/concepts/design/TitleWithIcon';
-import { ProjectObjectType } from '@odh-dashboard/internal/concepts/design/utils';
+import { ProjectObjectType } from '@odh-dashboard/ui-core';
 import { mlflowExperimentsBaseRoute } from '@odh-dashboard/internal/routes/pipelines/mlflow';
 import MlflowExperimentsPage from './MlflowExperimentsPage';
 import { EXPERIMENTS_PAGE_TITLE, EXPERIMENTS_NO_PROJECTS_MESSAGE } from '../shared/const';
