+++ packages/model-serving-backport/extensions.ts
@@ -1,6 +1,5 @@
 import { SupportedArea } from '@odh-dashboard/plugin-core/areas';
-// eslint-disable-next-line no-restricted-syntax
-import { ProjectObjectType } from '@odh-dashboard/internal/concepts/design/utils';
+import type { ProjectObjectType } from '@odh-dashboard/ui-core';
 // eslint-disable-next-line no-restricted-syntax
 import { NamespaceApplicationCase } from '@odh-dashboard/internal/pages/projects/types';
 import type { ModelServingPlatformExtension } from '@odh-dashboard/model-serving/extension-points';
