+++ packages/model-serving/src/components/overview/NoProjectServingEnabledSection.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 import { CardBody, FlexItem, Content } from '@patternfly/react-core';
-import { ProjectObjectType, SectionType } from '@odh-dashboard/internal/concepts/design/utils';
+import { ProjectObjectType, SectionType, CollapsibleSection } from '@odh-dashboard/ui-core';
 import OverviewCard from '@odh-dashboard/internal/pages/projects/screens/detail/overview/components/OverviewCard';
-import CollapsibleSection from '@odh-dashboard/internal/concepts/design/CollapsibleSection';
 
 const EmptyModelServingPlatformSection: React.FC = () => (
   <CollapsibleSection title="Serve models" data-testid="section-model-server">
