+++ packages/ui-core/src/design/TypeBorderedCard.tsx
@@ -1,0 +1,40 @@
+import * as React from 'react';
+import { css } from '@patternfly/react-styles';
+import { Card, CardProps } from '@patternfly/react-core';
+import { ProjectObjectType, SectionType } from './types';
+
+import './TypeBorderCard.scss';
+
+type TypeBorderedCardProps = CardProps & {
+  objectType?: ProjectObjectType;
+  sectionType?: SectionType;
+  selectable?: boolean;
+  selected?: boolean;
+  'data-testid'?: string;
+};
+const TypeBorderedCard: React.FC<TypeBorderedCardProps> = ({
+  objectType,
+  sectionType,
+  className,
+  selectable,
+  selected,
+  'data-testid': dataTestId,
+  ...rest
+}) => (
+  <Card
+    data-testid={dataTestId}
+    className={css(
+      className,
+      'odh-type-bordered-card',
+      sectionType,
+      objectType,
+      selectable && 'm-is-selectable',
+      selected && 'm-is-selected',
+    )}
+    isClickable={selectable}
+    isClicked={selected}
+    {...rest}
+  />
+);
+
+export default TypeBorderedCard;
