+++ packages/ui-core/src/design/index.ts
@@ -1,0 +1,11 @@
+export { ProjectObjectType, SectionType } from './types';
+export {
+  typedIconColor,
+  typedBackgroundColor,
+  typedColor,
+  sectionTypeIconColor,
+  sectionTypeBackgroundColor,
+  sectionTypeBorderColor,
+  sectionTypeLabelColor,
+} from './colorUtils';
+export { default as TypeBorderedCard } from './TypeBorderedCard';
