+++ packages/ui-core/src/design/types.ts
@@ -1,0 +1,57 @@
+export enum SectionType {
+  setup = 'set-up',
+  organize = 'organize',
+  training = 'training',
+  serving = 'serving',
+  general = 'general',
+}
+
+export enum ProjectObjectType {
+  project = 'project',
+  projectContext = 'projectContext',
+  notebook = 'notebook',
+  notebookImage = 'notebookImage',
+  build = 'build',
+  pipelineSetup = 'pipeline-setup',
+  pipeline = 'pipeline',
+  pipelineRun = 'pipeline-run',
+  pipelineExperiment = 'pipeline-experiment',
+  pipelineExecution = 'pipeline-execution',
+  pipelineArtifact = 'pipeline-artifact',
+  modelCustomization = 'model-customization',
+  modelEvaluation = 'model-evaluation',
+  labTuning = 'lab-tuning',
+  clusterStorage = 'cluster-storage',
+  model = 'model',
+  singleModel = 'single-model',
+  multiModel = 'multi-model',
+  modelServer = 'model-server',
+  modelCatalog = 'model-catalog',
+  mlflow = 'mlflow',
+  registeredModels = 'registered-models',
+  modelRegistryContext = 'modelRegistryContext',
+  deployedModels = 'deployed-models',
+  deployingModels = 'deploying-models',
+  connectedModels = 'connected-models',
+  modelRegistrySettings = 'model-registry-settings',
+  modelRegistry = 'model-registry',
+  servingRuntime = 'serving-runtime',
+  distributedWorkload = 'distributed-workload',
+  dataConnection = 'data-connection',
+  connections = 'connections',
+  clusterSettings = 'cluster-settings',
+  hardwareProfile = 'hardware-profile',
+  permissions = 'permissions',
+  user = 'user',
+  group = 'group',
+  storageClasses = 'storageClasses',
+  enabledApplications = 'enabled-applications',
+  exploreApplications = 'explore-applications',
+  resources = 'resources',
+  featureStore = 'feature-store',
+  promptManagement = 'prompt-management',
+  mcpCatalog = 'mcp-catalog',
+  agentOps = 'agent-ops',
+  taskAssistant = 'task-assistant',
+  apiKeys = 'api-keys',
+}
