+++ packages/ui-core/src/index.ts
@@ -15,3 +15,11 @@ export { default as DashboardPopupIconButton } from './components/DashboardPopup
 export { default as StateActionToggle } from './components/StateActionToggle';
 export type { ToggleState, StateActionToggleProps } from './components/StateActionToggle';
 export { default as DashboardEmptyTableView } from './components/DashboardEmptyTableView';
+
+export { default as CollapsibleSection } from './components/CollapsibleSection';
+export { default as EmptyDetailsView } from './components/EmptyDetailsView';
+
+export { default as PopoverListContent } from './components/PopoverListContent';
+export { default as WhosMyAdministrator } from './components/WhosMyAdministrator';
+
+export * from './design';
